Change Vulcanizing Agents for More Realism
Historically, Vulanization used mostly Sulfur with Carbon (graphite), Kaolinite, and Chalk as fillers. None of the fillers helped speed up the process of Vulcanization (which is very slow with just Sulfur), Zinc Oxide was needed to increase the speed.

This change modifies the recipes to make Sulfur as the main ingredient in the agent, and also make the player use Zinc Dust (zinc oxide) as well. The Zinc Dust can be easily obtained once the player has reached the Crusher and Macerator as those recipes already exist. This also adds recipes for the Quern that are not good as far as the amount of return of Zinc Dust.

I also wanted to make the recipe more flexible. Right now, the player has no chance at Rubber without Chalk. That is not realistic and is too limiting of a recipe. Historically, chalk was not required, but it was used alongside the Sulfur when available. There now exists a recipe that uses 2x Sulfur Dust, 1 Zinc Dust, 1 Kaolinite Dust and 1 Graphite Dust and an alternative recipe that uses 2x Sulfur Dust, 1 Zinc Dust, 1 Kaolinite Dust and 1 Chalk Powder.

To make it even more realistic (but probably not as good gameplay-wise), ONLY 1 Kaolinite or Graphite or Chalk should be used.